requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
airtable-python-wrapper=0.13.0=pypi_0
alembic=1.4.2=pypi_0
backcall=0.2.0=py_0
bandit=1.6.2=pypi_0
certifi=2020.6.20=py36_0
cffi=1.14.0=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
colorama=0.4.3=pypi_0
cryptography=2.9.2=pypi_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=py_0
facenet-pytorch=2.5.2=pypi_0
flask=1.1.2=pypi_0
flask-login=0.5.0=pypi_0
flask-migrate=2.5.3=pypi_0
flask-sqlalchemy=2.4.3=pypi_0
flask-wtf=0.14.3=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.3=pypi_0
html5lib=1.1=pypi_0
idna=2.9=pypi_0
imageio=2.10.5=pypi_0
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
itsdangerous=1.1.0=pypi_0
jedi=0.17.2=py36_0
jinja2=2.11.2=pypi_0
joblib=1.1.0=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py36_0
kiwisolver=1.3.1=pypi_0
libsodium=1.0.18=h62dcd97_0
lxml=4.5.1=pypi_0
mako=1.1.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=pypi_0
ndg-httpsclient=0.5.1=pypi_0
networkx=2.5.1=pypi_0
numpy=1.19.0=pypi_0
opencv-python=4.5.4.58=pypi_0
pandas=1.0.5=pypi_0
parso=0.7.0=py_0
pbr=5.4.5=pypi_0
pickleshare=0.7.5=py36_0
pillow=8.4.0=pypi_0
pip=20.1.1=py_1
prompt-toolkit=3.0.8=py_0
pusher=3.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.1=py_0
pynacl=1.4.0=pypi_0
pyopenssl=19.1.0=pypi_0
pyparsing=3.0.6=pypi_0
python=3.6.10=he025d50_1009_cpython
python-dateutil=2.8.1=py_0
python-editor=1.0.4=pypi_0
python_abi=3.6=1_cp36m
pytz=2020.1=pypi_0
pywavelets=1.1.1=pypi_0
pywin32=227=py36he774522_1
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py36ha925a31_1
requests=2.23.0=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=47.1.1=py36h9f0ad1d_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlalchemy=1.3.17=pypi_0
stevedore=2.0.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2020.9.3=pypi_0
torch=1.10.0=pypi_0
torchvision=0.11.1=pypi_0
tornado=6.0.4=py36he774522_1
traitlets=4.3.3=py36_0
typing-extensions=4.0.0=pypi_0
urllib3=1.25.9=pypi_0
vc=14.1=h869be7e_1
vs2015_runtime=14.16.27012=h30e32a0_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py_1
wincertstore=0.2=py36_1003
wtforms=2.3.1=pypi_0
zeromq=4.3.2=ha925a31_3
